endchan: Implement IP block bypass
Endchan lets users bypass IP block in some circumstances. If a user's IP is blocked and the bypass is possible, a dialog will pop up, the user will need to enter a captcha, and then he will be able to post until the bypass expires. Then the user will need to repeat the process.